# docs/developers/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/templates": {
      "post": {
        "operationId": "templateCreate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "envelopeId": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 1000
                  },
                  "roleNames": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 200
                    }
                  }
                },
                "required": [
                  "envelopeId",
                  "name"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The template, with its roles and fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateDetail"
                }
              }
            }
          },
          "401": {
            "description": "Error codes: API_KEY_INVALID.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Error codes: API_KEY_READ_ONLY.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Error codes: NOT_FOUND.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Error codes: TEMPLATE_NAME_TAKEN.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "410": {
            "description": "Error codes: ENVELOPE_PURGED.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Error codes: NOT_READY_TO_SEND.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Error codes: RATE_LIMITED.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "session": []
          }
        ],
        "summary": "Save an envelope as a template",
        "tags": [
          "Templates"
        ],
        "description": "Turn a prepared envelope into a reusable template: its PDF, its people as named roles, and where they sign. The template keeps its own copy of the PDF.\n\n- envelopeId (required): an envelope in your workspace with at least one person and every signer given a required field. It may be a draft or already sent.\n\n- name (required): 1–120 characters, different from every other active template.\n\n- description (optional): up to 1,000 characters.\n\n- roleNames (optional): an object from recipient id to the role’s name, such as \"Patient\". A role is named after the person on the envelope unless listed here. Names must differ.",
        "x-idempotency": "none",
        "x-rate-limit": "30 per 60s per workspace"
      },
      "get": {
        "operationId": "templateList",
        "parameters": [
          {
            "name": "archived",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "true",
                "false"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Template summaries, at most 500.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateList"
                }
              }
            }
          },
          "401": {
            "description": "Error codes: API_KEY_INVALID.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Error codes: RATE_LIMITED.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "session": []
          }
        ],
        "summary": "List templates",
        "tags": [
          "Templates"
        ],
        "description": "Active templates, newest first. Use an id from here to create envelopes from it.\n\n- archived (optional query): true lists archived templates instead. Defaults to false.",
        "x-idempotency": "none"
      }
    },
    "/api/v1/templates/{id}": {
      "get": {
        "operationId": "templateGet",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The template with its roles and fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateDetail"
                }
              }
            }
          },
          "401": {
            "description": "Error codes: API_KEY_INVALID.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Error codes: TEMPLATE_NOT_FOUND.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Error codes: RATE_LIMITED.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "session": []
          }
        ],
        "summary": "Read a template",
        "tags": [
          "Templates"
        ],
        "description": "One template with its roles and fields. An archived template can still be read, but cannot start new envelopes.\n\n- id (path): the template id.",
        "x-idempotency": "none"
      },
      "patch": {
        "operationId": "templateUpdate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "description": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 1000
                  },
                  "defaultMessage": {
                    "nullable": true,
                    "type": "string",
                    "maxLength": 2000
                  },
                  "archived": {
                    "type": "boolean"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated template.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateDetail"
                }
              }
            }
          },
          "401": {
            "description": "Error codes: API_KEY_INVALID.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Error codes: API_KEY_READ_ONLY.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Error codes: TEMPLATE_NOT_FOUND.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Error codes: TEMPLATE_NAME_TAKEN.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Error codes: RATE_LIMITED.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "session": []
          }
        ],
        "summary": "Rename, archive or restore a template",
        "tags": [
          "Templates"
        ],
        "description": "Change a template’s name, description, default message or whether it is archived. The layout cannot be edited: save a new template instead.\n\n- name (optional): 1–120 characters, different from every other active template.\n\n- description and defaultMessage (optional): text, or null to clear.\n\n- archived (optional): true hides the template from new use, false restores it.\n\n- Send at least one of these.",
        "x-idempotency": "none",
        "x-rate-limit": "30 per 60s per workspace"
      }
    },
    "/api/v1/templates/{id}/envelopes": {
      "post": {
        "description": "Give each role of the template a name and an email, and get a draft (or a sent envelope) with the people, fields and signing order already in place. It is an ordinary envelope from then on.\n\n- recipients (required): one entry per role of the template, each `{ role, name, email }`, where `role` is the template role’s name (for example \"Patient\"). No role may be missing or repeated, and no email may be used twice.\n\n- send (optional): true sends it at once, false (the default) leaves a draft you can still edit.\n\n- message (optional): the note in the invitation; defaults to the template’s. title (optional): defaults to the template’s name.\n\n- externalId and metadata (optional): as on upload.\n\n- Idempotency-Key (optional): as on upload. Repeating the same key and body within 24 hours returns the envelope the first request created (Idempotency-Replayed: true).",
        "operationId": "templateEnvelope",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Optional. A value unique to this attempt (a UUID is ideal). Repeating the request with the same key and the same body within 24 hours returns the resource the first request created (with Idempotency-Replayed: true) instead of creating another. Without a key, every request creates a new one.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "recipients": {
                    "minItems": 1,
                    "maxItems": 50,
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "role": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 200
                        },
                        "name": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 200
                        },
                        "email": {
                          "type": "string",
                          "maxLength": 254
                        }
                      },
                      "required": [
                        "role",
                        "name",
                        "email"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "message": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "externalId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200,
                    "pattern": "^[a-zA-Z0-9_.:@-]+$"
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string",
                      "maxLength": 500
                    }
                  },
                  "send": {
                    "default": false,
                    "type": "boolean"
                  }
                },
                "required": [
                  "recipients"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Envelope detail (excerpt). Policy is frozen when this call runs, not when the template was saved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeDetail"
                }
              }
            }
          },
          "401": {
            "description": "Error codes: API_KEY_INVALID.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Error codes: API_KEY_READ_ONLY.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Error codes: TEMPLATE_NOT_FOUND.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Error codes: TEMPLATE_ARCHIVED.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Error codes: TEMPLATE_ROLE_MISMATCH, DOCUMENT_CATEGORY_BLOCKED, IDEMPOTENCY_KEY_MISMATCH.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Error codes: RATE_LIMITED.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "session": []
          }
        ],
        "summary": "Create an envelope from a template",
        "tags": [
          "Templates"
        ],
        "x-idempotency": "optional",
        "x-rate-limit": "100 per 60s per workspace"
      }
    },
...
  },
  "info": {
    "title": "Envelope powered by HealthProHub API",
    "description": "Server-to-server API for creating, sending and tracking documents for signature.\n\n- **Base URL:** `/api/v1` on your Envelope host.\n- **Authentication:** `Authorization: Bearer <API key>` (scheme `apiKey`). A key can call only the operations listed here; every other route needs a signed-in session.\n- **Errors:** RFC 7807 `application/problem+json` with a stable `code`.\n- **Webhooks:** see the `x-webhooks` extension and the `Webhook*` schemas.\n- **Guide:** `docs/developers/` in the repository is the same contract as a readable guide.",
    "version": "0.9.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "A server API key, created in Settings → Integrations. Keep it on your server."
      },
      "session": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "A signed-in person’s session token (the web app)."
      },
      "embedSession": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "The token an embedded editor session receives after its launch handshake."
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "code": {
            "type": "string",
            "enum": [
              "BAD_REQUEST",
              "VALIDATION_FAILED",
              "UNAUTHENTICATED",
              "FORBIDDEN",
              "NOT_FOUND",
              "CONFLICT",
              "PAYLOAD_TOO_LARGE",
              "UNSUPPORTED_MEDIA_TYPE",
              "RATE_LIMITED",
              "INTERNAL_ERROR",
              "SERVICE_UNAVAILABLE",
              "EMBED_ORIGIN_NOT_ALLOWED",
              "EMBED_SESSION_INVALID",
              "EMBED_SESSION_EXPIRED",
              "EMBED_LAUNCH_USED",
              "EMBED_SCOPE_DENIED",
              "EMBED_UPLOAD_BOUND",
              "INVALID_CREDENTIALS",
              "SESSION_EXPIRED",
              "EMAIL_ALREADY_REGISTERED",
              "FORBIDDEN_ROLE",
              "LAST_OWNER",
              "INVITE_TOKEN_INVALID",
              "INVITE_TOKEN_EXPIRED",
              "PASSWORD_RESET_TOKEN_INVALID",
              "PASSWORD_RESET_TOKEN_EXPIRED",
              "CURRENT_PASSWORD_INCORRECT",
              "TWO_FACTOR_CODE_INVALID",
              "TWO_FACTOR_CHALLENGE_INVALID",
              "TWO_FACTOR_REQUIRED",
              "TWO_FACTOR_ALREADY_ENABLED",
              "TWO_FACTOR_NOT_ENABLED",
              "FILE_REQUIRED",
              "FILE_TOO_LARGE",
              "UNSUPPORTED_FILE_TYPE",
              "INVALID_PDF",
              "ENCRYPTED_PDF",
              "PAGE_LIMIT_EXCEEDED",
              "MALWARE_DETECTED",
              "INVALID_COORDINATE_SPACE",
              "RATIO_OUT_OF_RANGE",
              "FIELD_EXCEEDS_PAGE",
              "PAGE_OUT_OF_RANGE",
              "ENVELOPE_NOT_DRAFT",
              "ENVELOPE_TERMINAL",
              "ENVELOPE_EXPIRED",
              "RECIPIENT_EMAIL_TAKEN",
              "DRAFT_REVISION_MISMATCH",
              "DOCUMENT_CATEGORY_BLOCKED",
              "RECIPIENT_HAS_NO_FIELDS",
              "NOT_READY_TO_SEND",
              "REQUIRED_FIELDS_INCOMPLETE",
              "TOKEN_INVALID",
              "TOKEN_EXPIRED",
              "TOKEN_ALREADY_USED",
              "CONSENT_REQUIRED",
              "CONSENT_TEXT_CHANGED",
              "INVALID_SIGNATURE_IMAGE",
              "REMINDER_TOO_SOON",
              "DOWNLOAD_LINK_EXPIRED",
              "DOWNLOAD_RENEW_TOO_SOON",
              "ENVELOPE_ON_LEGAL_HOLD",
              "ENVELOPE_PURGED",
              "TEMPLATE_NOT_FOUND",
              "TEMPLATE_ARCHIVED",
              "TEMPLATE_NAME_TAKEN",
              "TEMPLATE_ROLE_MISMATCH",
              "BULK_TOO_LARGE",
              "BULK_BATCH_NOT_FOUND",
              "API_KEY_INVALID",
              "API_KEY_NOT_ALLOWED",
              "API_KEY_READ_ONLY",
              "WEBHOOK_URL_NOT_ALLOWED",
              "WEBHOOK_ENDPOINT_LIMIT_REACHED",
              "WEBHOOK_ENDPOINT_TOTAL_LIMIT_REACHED",
              "WEBHOOK_ENDPOINT_ACTIVE",
              "WEBHOOK_DELIVERY_NOT_REDRIVABLE",
              "IDEMPOTENCY_KEY_REQUIRED",
              "IDEMPOTENCY_KEY_MISMATCH"
            ]
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "requestId": {
            "description": "Matches the X-Request-Id header.",
            "type": "string"
          },
          "errors": {
            "description": "Per-field problems, on VALIDATION_FAILED and NOT_READY_TO_SEND.",
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ],
              "additionalProperties": {}
            }
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "additionalProperties": {},
        "description": "RFC 7807 problem details (application/problem+json)."
      },
...
      "TemplateRole": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "SIGNER",
              "APPROVER",
              "VIEWER",
              "CC"
            ]
          },
          "routingOrder": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "colorIndex": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "id",
          "name",
          "role",
          "routingOrder",
          "colorIndex"
        ],
        "additionalProperties": {}
      },
      "TemplateField": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "templateRoleId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "SIGNATURE",
              "INITIALS",
              "DATE_SIGNED",
              "TEXT_INPUT",
              "CHECKBOX"
            ]
          },
          "pageNumber": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "ratioX": {
            "type": "number"
          },
          "ratioY": {
            "type": "number"
          },
          "ratioWidth": {
            "type": "number"
          },
          "ratioHeight": {
            "type": "number"
          },
          "required": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "templateRoleId",
          "type",
          "pageNumber",
          "ratioX",
          "ratioY",
          "ratioWidth",
          "ratioHeight",
          "required"
        ],
        "additionalProperties": {}
      },
...
      "TemplateDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "nullable": true,
            "type": "string"
          },
          "pageCount": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "documentCategory": {
            "type": "string"
          },
          "roleCount": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "fieldCount": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "archivedAt": {
            "nullable": true,
            "type": "string",
            "description": "UTC ISO 8601 timestamp",
            "example": "2026-09-27T10:00:00.000Z"
          },
          "createdAt": {
            "type": "string",
            "description": "UTC ISO 8601 timestamp",
            "example": "2026-09-27T10:00:00.000Z"
          },
          "createdByName": {
            "type": "string"
          },
          "defaultMessage": {
            "nullable": true,
            "type": "string"
          },
          "sequentialSigning": {
            "type": "boolean"
          },
          "reminderIntervalDays": {
            "nullable": true,
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TemplateRole"
            }
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TemplateField"
            }
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "pageCount",
          "documentCategory",
          "roleCount",
          "fieldCount",
          "archivedAt",
          "createdAt",
          "createdByName",
          "defaultMessage",
          "sequentialSigning",
          "reminderIntervalDays",
          "roles",
          "fields"
        ],
        "additionalProperties": {}
      },
      "TemplateList": {
        "type": "object",
        "properties": {
          "templates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TemplateSummary"
            }
          }
        },
        "required": [
          "templates"
        ],
        "additionalProperties": {}
      },
...
    }
  },
...
}